Move ControlledBloq to `_infra` (#587)
