+++ qualtran/__init__.py
@@ -62,6 +62,8 @@ from ._infra.gate_with_registers import GateWithRegisters
 
 from ._infra.adjoint import Adjoint
 
+from ._infra.controlled import ControlledBloq
+
 from ._infra.bloq_example import BloqExample, bloq_example, BloqDocSpec
 
 # --------------------------------------------------------------------------------------------------
